fix: introduced puppeteer instead of puppeteer core
the backend was deployed on vercel, where I have to manage the node version 14.0.0. In the recent release, vercel suggested to use minimum 16.0.0 version. This version is not suitable to run puppeteer core thats why the backend is now shifted to render.com. Also, we dockerized the project so that we can use puppeteer instead of puppeteer core on production by fixing chrome related issues by docker.